+++ server/api/core/rightJoin.js
@@ -1,0 +1,74 @@
+/**
+ * @summary Returns an disjoint object as right join. For a visualization, see:
+ *          http://www.codeproject.com/KB/database/Visual_SQL_Joins/Visual_SQL_JOINS_orig.jpg
+ *          Additionally, the join is done recursively on properties of
+ *          nested objects as well. Nested arrays are handled like
+ *          primitive values.
+ * @author Tom De Caluwé
+ * @param {Object} leftSet An object that can contain nested sub-objects
+ * @param {Object} rightSet An object that can contain nested sub-objects
+ * @returns {Object} The disjoint object that does only contain properties
+ *                   from the rightSet. But only those, that were not present
+ *                   in the leftSet.
+ */
+const doRightJoinNoIntersection = (leftSet, rightSet) => {
+  if (rightSet === null) return null;
+
+  let rightJoin;
+  if (Array.isArray(rightSet)) {
+    rightJoin = [];
+  } else {
+    rightJoin = {};
+  }
+  const findRightOnlyProperties = () => {
+    return Object.keys(rightSet).filter(function (key) {
+      if (typeof(rightSet[key]) === "object" &&
+        !Array.isArray(rightSet[key])) {
+        // Nested objects are always considered
+        return true;
+      }
+      // Array or primitive value
+      return !leftSet.hasOwnProperty(key);
+    });
+  };
+
+  for (const key of findRightOnlyProperties()) {
+    if (typeof(rightSet[key]) === "object") {
+      // subobject or array
+      if (leftSet.hasOwnProperty(key) && (typeof(leftSet[key]) !== "object" ||
+           Array.isArray(leftSet[key]) !== Array.isArray(rightSet[ key ]))) {
+        // This is not expected!
+        throw new Error(
+          "Left object and right object's internal structure must be " +
+          "congruent! Offending key: " + key
+        );
+      }
+      const rightSubJoin = doRightJoinNoIntersection(
+        leftSet.hasOwnProperty(key) ? leftSet[key] : {},
+        rightSet[key]
+      );
+
+      const obj = {};
+      if (rightSubJoin === null) {
+        obj[key] = null;
+      } else if (Object.keys(rightSubJoin).length !== 0 ||
+                 Array.isArray(rightSubJoin)) {
+        // object or (empty) array
+        obj[key] = rightSubJoin;
+      }
+      rightJoin = Object.assign(rightJoin, obj);
+    } else {
+      // primitive value (or array)
+      if (Array.isArray(rightSet)) {
+        rightJoin.push(rightSet[key]);
+      } else {
+        const obj = {};
+        obj[key] = rightSet[key];
+        rightJoin = Object.assign(rightJoin, obj);
+      }
+    }
+  }
+  return rightJoin;
+};
+
+export default doRightJoinNoIntersection;
